refactor: extract FastAPI app factory from main module
